requirements.txt
aws-cdk-lib>=2.200.0
constructs>=10.0.0
aws-cdk.aws-neptune-alpha>=2.185.0a0
boto3>=1.34.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
shap>=0.42.0
gremlinpython>=3.6.0
